requirements.txt
# Runtime dependencies
fastapi>=0.121.2
fastmcp>=2.13.1
httpx>=0.28.1
pydantic>=2.12.4
pydantic-settings>=2.12.0
uvicorn[standard]>=0.38.0
sentence-transformers>=5.1.2
faiss-cpu>=1.13.0
chromadb>=0.5.3
python-dotenv>=1.0.1
pyyaml>=6.0.2
openai>=1.52.0
google-generativeai>=0.8.0
requests>=2.32.3
beautifulsoup4>=4.12.3
tqdm>=4.66.5
markdownify>=0.13.1

# Development / testing
pytest>=9.0.1
pytest-asyncio>=1.3.0
ruff>=0.14.5
mypy>=1.13
einops>=0.8.1

PyGithub>=2.3.0
tiktoken>=0.12.0
google-genai>=1.51.0
